+++ src/app/admin/gigs/_gig-form/EditGigFormClient.tsx
@@ -8,16 +8,16 @@ import { toast } from '@/hooks/use-toast';
 import { Card, CardHeader, CardTitle, CardDescription, CardContent } from '@/components/ui/card';
 import type { Country } from '@/lib/countries.server';
 import { useRouter } from 'next/navigation';
-import GigFormFields from '@/app/gig-form/_components/GigFormFields';
-import PosterField from '@/app/gig-form/_components/PosterField';
+import GigFormFields from '@/app/admin/gigs/_gig-form/_components/GigFormFields';
+import PosterField from '@/app/admin/gigs/_gig-form/_components/PosterField';
 import type { GigUpsertResponse } from '@/lib/gig-form-api';
 import { updateGig } from '@/lib/gig-form-api';
-import { defaultGigFormValues, gigFormSchema } from '@/app/gig-form/gig-form.shared';
-import type { GigFormValues } from '@/app/gig-form/gig-form.shared';
-import { useEditGigFormData } from '@/app/gig-form/useEditGigFormData';
-import { useGigLookup } from '@/app/gig-form/useGigLookup';
-import { useGigSubmit } from '@/app/gig-form/useGigSubmit';
-import { buildGigFormPublicIdPath, ADMIN_GIGS_BASE_PATH } from '@/app/gig-form/gig-form-paths';
+import { defaultGigFormValues, gigFormSchema } from '@/app/admin/gigs/_gig-form/gig-form.shared';
+import type { GigFormValues } from '@/app/admin/gigs/_gig-form/gig-form.shared';
+import { useEditGigFormData } from '@/app/admin/gigs/_gig-form/useEditGigFormData';
+import { useGigLookup } from '@/app/admin/gigs/_gig-form/useGigLookup';
+import { useGigSubmit } from '@/app/admin/gigs/_gig-form/useGigSubmit';
+import { buildAdminGigPublicIdPath, ADMIN_GIGS_BASE_PATH } from '@/app/admin/gigs/admin-gig-paths';
 
 interface EditGigFormClientProps {
   readonly countries: Country[];
@@ -49,7 +49,7 @@ export default function EditGigFormClient(props: EditGigFormClientProps) {
         description: 'Gig was updated.',
       });
       const returnHref =
-        result.publicId && buildGigFormPublicIdPath(ADMIN_GIGS_BASE_PATH, result.publicId);
+        result.publicId && buildAdminGigPublicIdPath(ADMIN_GIGS_BASE_PATH, result.publicId);
       if (returnHref) {
         router.push(returnHref);
       } else {
