+++ src/app/admin/gigs/_gig-form/useEditGigFormData.test.ts
@@ -1,15 +1,15 @@
 // @vitest-environment jsdom
 
 import { act, renderHook, waitFor } from '@testing-library/react';
+import type { UseFormReturn } from 'react-hook-form';
 import { useForm } from 'react-hook-form';
 
 import type { AdminGigFormData } from '@/app/admin/gigs/types';
-import type { GigFormValues } from '@/app/gig-form/gig-form.shared';
-import type { UseFormReturn } from 'react-hook-form';
-
-import { defaultGigFormValues } from '@/app/gig-form/gig-form.shared';
+import { GigStatusAPI } from '@/app/admin/gigs/types';
+import type { GigFormValues } from '@/app/admin/gigs/_gig-form/gig-form.shared';
+import { defaultGigFormValues } from '@/app/admin/gigs/_gig-form/gig-form.shared';
 import { fetchAdminGigByPublicId } from '@/lib/admin-api';
-import { useEditGigFormData } from '@/app/gig-form/useEditGigFormData';
+import { useEditGigFormData } from '@/app/admin/gigs/_gig-form/useEditGigFormData';
 import { createQueryClientWrapper, createTestQueryClient } from '@/test-utils/react-query-client';
 
 const { toastMock } = vi.hoisted(() => ({
@@ -45,7 +45,7 @@ function createAdminGigFormData(overrides: Partial<AdminGigFormData> = {}): Admi
   return {
     publicId: 'gig-public-id',
     title: 'Arctic Monkeys',
-    status: 'Pending',
+    status: GigStatusAPI.Pending,
     date: '2026-07-01T20:00:00.000Z',
     endDate: '2026-07-02T22:00:00.000Z',
     city: 'Barcelona',
