+++ src/app/suggest/suggest-launch.ts
@@ -1,4 +1,4 @@
-import { buildGigFormEditPath, ADMIN_GIGS_BASE_PATH } from '@/app/gig-form/gig-form-paths';
+import { buildAdminGigEditPath, ADMIN_GIGS_BASE_PATH } from '@/app/admin/gigs/admin-gig-paths';
 
 const GIG_FORM_START_PARAM_PATTERN = /^[a-z0-9-]{1,64}$/i;
 
@@ -8,7 +8,7 @@ export function resolveAdminGigLaunchPath(startParam: string | undefined): strin
     return `${ADMIN_GIGS_BASE_PATH}/new`;
   }
 
-  return buildGigFormEditPath(ADMIN_GIGS_BASE_PATH, trimmedStartParam);
+  return buildAdminGigEditPath(ADMIN_GIGS_BASE_PATH, trimmedStartParam);
 }
 
 export function resolveSuggestLaunchPath(isAdmin: boolean, startParam: string | undefined): string {
